# backend/docs/swagger.yaml
definitions:
...
  github_com_movie-matcher_backend_internal_services.SemanticSearchItem:
    properties:
      description:
        type: string
      media_id:
        type: integer
      media_type:
        type: string
      poster:
        type: string
      score:
        type: number
      title:
        type: string
    type: object
  github_com_movie-matcher_backend_internal_services.SemanticSearchResponse:
    properties:
      results:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_services.SemanticSearchItem'
        type: array
    type: object
...
  internal_handlers.CalendarRelease:
    properties:
      id:
        type: integer
      listStatus:
        type: string
      mediaType:
        type: string
      poster:
        type: string
      releaseDate:
        description: YYYY-MM-DD
        type: string
      title:
        type: string
      titleI18n: {}
    type: object
...
  internal_handlers.PrecomputedSimilarResponse:
    properties:
      crossType:
        description: среди других типов
        items:
          $ref: '#/definitions/internal_handlers.SimilarItem'
        type: array
      entityId:
        type: integer
      entityType:
        type: string
      sameType:
        description: среди своего типа (content + users)
        items:
          $ref: '#/definitions/internal_handlers.SimilarItem'
        type: array
    type: object
...
  internal_handlers.SimilarItem:
    properties:
      sameType:
        type: boolean
      score:
        type: number
      similarEntityId:
        type: integer
      similarEntityType:
        type: string
      source:
        description: '"content" | "users"'
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /calendar/releases:
    get:
      parameters:
      - description: From date YYYY-MM-DD
        in: query
        name: from
        required: true
        type: string
      - description: To date YYYY-MM-DD
        in: query
        name: to
        required: true
        type: string
      - default: movie
        description: movie|tv-series|anime|...
        in: query
        name: mediaType
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handlers.CalendarRelease'
            type: array
      summary: Get calendar releases
      tags:
      - Calendar
...
  /search/semantic:
    get:
      description: Search media by semantic similarity to the query (description,
        themes, mood). Requires recommendation service.
      parameters:
      - description: Search query (by meaning)
        in: query
        name: q
        required: true
        type: string
      - description: 'Filter: movie, anime, game, etc.'
        in: query
        name: mediaType
        type: string
      - description: Max results (default 20)
        in: query
        name: limit
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_movie-matcher_backend_internal_services.SemanticSearchResponse'
      summary: Semantic search by description/meaning
      tags:
      - search
  /similar/store/{entityType}/{id}:
    get:
      parameters:
      - description: Entity type
        in: path
        name: entityType
        required: true
        type: string
      - description: Entity ID
        in: path
        name: id
        required: true
        type: integer
      - description: content|users
        in: query
        name: source
        type: string
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.PrecomputedSimilarResponse'
      summary: Get precomputed similar entities
      tags:
      - Similar
...
swagger: "2.0"